Use C++20 (fixes #6552)
src/engine/client/client.h:137:49: warning: arithmetic between different enumeration types ('(unnamed enum at /Users/deen/git/ddnet/src/engine/shared/protocol.h:77:1)' and '(unnamed enum at /Users/deen/git/ddnet/src/base/system.h:838:1)') is deprecated [-Wdeprecated-anon-enum-enum-conversion]

src/engine/client/updater.cpp:297:2: error: use range-based for loop instead [modernize-loop-convert,-warnings-as-errors]

src/engine/client/serverbrowser.cpp:1003:21: error: use 'contains' to check for membership [readability-container-contains,-warnings-as-errors]

Requires some experiments first

+++ src/engine/client/client.h
@@ -134,7 +134,7 @@ class CClient : public IClient, public CDemoPlayer::IListener
 	CFriends m_Friends;
 	CFriends m_Foes;
 
-	char m_aConnectAddressStr[MAX_SERVER_ADDRESSES * NETADDR_MAXSTRSIZE];
+	char m_aConnectAddressStr[(size_t)MAX_SERVER_ADDRESSES * (size_t)NETADDR_MAXSTRSIZE];
 
 	CUuid m_ConnectionID;
